requirements.txt
# install torch manually based on your specifications
# torch version: 2.2.1

matplotlib~=3.8.3
pandas~=2.2.1
pillow~=10.2.0
tqdm~=4.66.2
opencv-python~=4.9.0.80
numpy~=1.26.4
yacs~=0.1.8
jupyter~=1.1.1